Fix STRL generation bug in DAG-Aware TetriSched.
Our prior DAG-aware STRL generation algorithm would silently drop
AllocationExpression STRLs if they coincided with other children that
could not lead to any Choose/WindowedChoose expressions.

Currently, we fix this by forcing all the generated
AllocationExpressions to be children of the root ObjectiveExpression.

This is not entirely correct since AllocationExpressions may still have
other parents that do not show up in a DFS/BFS from the root, but may
hinder the application of certain bottom-up optimization passes. We
leave this for future fixing, if the need arises.